Codes refactoring
